+++ libs/design-system/buttons/icon-button/src/lib/icon-button.module.ts
