+++ libs/design-system/nav-header/src/lib/nav-header.component.ts
@@ -4,7 +4,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { AppNavButtonComponent } from '@hra-ui/design-system/buttons/app-nav-button';
-import { IconButtonSizeDirective } from '@hra-ui/design-system/icon-button';
+import { IconButtonSizeDirective } from '@hra-ui/design-system/buttons/icon-button';
 import { SoftwareStatus } from '@hra-ui/design-system/indicators/software-status-indicator';
 import { NavHeaderButtonsComponent, NavHeaderButtonsVariant } from '@hra-ui/design-system/nav-header-buttons';
 import { ScrollingModule } from '@hra-ui/design-system/scrolling';
